+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/e2e-tests
 
+## 2.8.27
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.27
+  - @stackframe/js@2.8.27
+
+## 2.8.26
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.26
+  - @stackframe/js@2.8.26
+
 ## 2.8.25
 
 ### Patch Changes