+++ apps/mock-oauth-server/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @stackframe/mock-oauth-server
 
+## 2.8.27
+
+## 2.8.26
+
 ## 2.8.25
 
 ## 2.8.24