+++ examples/docs-examples/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/docs-examples
 
+## 2.8.27
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.27
+  - @stackframe/stack@2.8.27
+  - @stackframe/stack-ui@2.8.27
+
+## 2.8.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.26
+  - @stackframe/stack@2.8.26
+  - @stackframe/stack-ui@2.8.26
+
 ## 2.8.25
 
 ### Patch Changes