+++ examples/react-example/CHANGELOG.md
@@ -1,5 +1,17 @@
 # react-example
 
+## 2.8.27
+
+### Patch Changes
+
+- @stackframe/react@2.8.27
+
+## 2.8.26
+
+### Patch Changes
+
+- @stackframe/react@2.8.26
+
 ## 2.8.25
 
 ### Patch Changes