+++ packages/init-stack/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/init-stack
 
+## 2.8.27
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.27
+
+## 2.8.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.26
+
 ## 2.8.25
 
 ### Patch Changes